+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add "org.bytedeco.javacpp.pathsfirst" system property to let users search "java.library.path", etc before the class path
  * Add `Parser` support for `_Bool`, `_Complex`, `_Imaginary`, `complex`, `imaginary` types from C99
  * Fix `Generator` incorrectly splitting type names for template arguments containing function types
  * Fix `NullPointerException` in `Builder` when copying resources for static libraries